+++ src/lib/constants.js
@@ -1,0 +1,1 @@
+export const ORDER_OPTIONS = ['VOTE', 'DATE', 'COMMENTS'];