+++ CHANGELOG.md
@@ -5,6 +5,9 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Added
+- Buttons for changing the order of the search results
+- Add web-ext config file for easier build toolchain
 
 ## [1.2.1] - 2019-02-18
 ### Added